+++ src/app/designer.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { DesignerProduct } from './designer-product';
 
@@ -15,4 +15,17 @@ export class DesignerService {
   deductProduct(product: DesignerProduct): Observable<DesignerProduct> {
     return this.http.put<DesignerProduct>(`${this.apiUrl}`, product);
   }
+
+  downloadFile(id: number, fileName: string): void {
+    this.http.get('http://localhost:8080/api/designer/download/' + id, {
+      responseType: 'arraybuffer',
+      headers: new HttpHeaders().append('Content-Type', 'application/json')
+    }).subscribe((response: ArrayBuffer) => {
+      const blob = new Blob([response], { type: 'application/octet-stream' });
+      const downloadLink = document.createElement('a');
+      downloadLink.href = window.URL.createObjectURL(blob);
+      downloadLink.download = fileName;
+      downloadLink.click();
+    });
+}
 }
